Bug 2028887 - Use 64-bit arithmetic.  a=diannaS DONTBUILD
(As suggested by clauditor.)

Original Revision: https://phabricator.services.mozilla.com/D291997

Differential Revision: https://phabricator.services.mozilla.com/D293453

+++ gfx/thebes/COLRFonts.cpp
@@ -2335,7 +2335,7 @@ bool ItemVariationData::Validate(const COLRv1Header* aHeader,
   uint32_t deltaSetSize =
       (uint16_t(regionIndexCount) + uint16_t(wordDeltaCount)) << longWords;
   if (reinterpret_cast<const char*>(deltaSets()) +
-          uint16_t(itemCount) * deltaSetSize >
+          uint64_t(uint16_t(itemCount)) * deltaSetSize >
       reinterpret_cast<const char*>(aHeader) + aLength) {
     return false;
   }
